Add authentication middleware
This is the initial addition of the middleware. It doesn't quite work as
required yet. But that's to be fixed in a subsequent commit. Here's what
it will do, in time:

Routes can be protected with the AuthenticationMiddleware. If it's
enabled for a given route, if the user is not authenticated, and makes a
GET request, they will be redirected to `/login`. From there, if they
make a POST request, the POST data will be retrieved from the
request and used to initialise the form. If the form validates, then
the authentication process will be run. If that succeeds then they will
be redirected back to the original route which they requested. If
authentication fails they'll see the login form again, where they will
also see the reason why the authentication failed (as much as possible).